Moved method to save compressed data to storage engine struct
